test(create-project): simplify the testings
